refactor(mesos): rewrite configuration system.
